refactor: remove db.py and reorganize database initialization
- Deleted db.py to streamline database management and reduce redundancy.
- Updated main.py to directly import init_db from the database module, simplifying the database initialization process.
- Introduced auth_router.py to handle authentication routes, enhancing modularity and organization of the codebase.
- Ensured dotenv loading is handled in the database module for consistent environment variable access.
