+++ src/backend/routers/auth_router.py
@@ -1,12 +1,11 @@
 import secrets
 import jwt
 import httpx
-from fastapi import APIRouter, Request, HTTPException, Depends
+from fastapi import APIRouter, Request, HTTPException
 from fastapi.responses import RedirectResponse, FileResponse, JSONResponse
 import os
 
 from config import get_auth_url, get_token_url, OIDC_CONFIG, set_session, delete_session, STATIC_DIR, get_session
-from dependencies import UserSession, require_auth
 from coder import CoderAPI
 
 auth_router = APIRouter()
